# openapi.yaml
openapi: "3.0.3"
info:
  title: Sistem Pengolahan Nilai Siswa API
  description: >
    API untuk aplikasi pengolahan nilai siswa.
    Mendukung 3 role: admin, guru, siswa.
    Formula: Nilai Akhir = (30% × Tugas) + (30% × UTS) + (40% × UAS).
    Syarat Lulus: Nilai Akhir ≥ 70.
  version: "1.0.0"
  contact:
    name: Developer
    email: dev@sekolah.id
servers:
  - url: http://localhost:8000/api
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token dari endpoint login
  schemas:
...
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Not found"
      required:
        - detail
    Message:
      type: object
      properties:
        message:
          type: string
          example: "Operasi berhasil"

    # ── Auth ──
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          example: "admin"
        password:
          type: string
          minLength: 1
          example: "admin123"
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIs..."
        user:
          $ref: '#/components/schemas/UserResponse'
      required:
        - token
        - user
    ChangePasswordRequest:
      type: object
      properties:
        old_password:
          type: string
          minLength: 1
          example: "admin123"
        new_password:
          type: string
          minLength: 6
          description: Minimal 6 karakter
          example: "newpass456"
      required:
        - old_password
        - new_password
    ResetPasswordRequest:
      type: object
      properties:
        new_password:
          type: string
          minLength: 6
          description: Minimal 6 karakter
          example: "resetpass789"
      required:
        - new_password

    # ── User ──
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin"
        role:
          type: string
          enum: [admin, guru, siswa]
          example: "admin"
        nama:
          type: string
          example: "Administrator"
      required:
        - id
        - username
        - role
        - nama
    UserCreate:
      type: object
      properties:
        username:
          type: string
          example: "guru_baru"
        password:
          type: string
          minLength: 6
          example: "password123"
        role:
          type: string
          enum: [admin, guru, siswa]
          example: "guru"
        nama:
          type: string
          example: "Nama Guru Baru"
      required:
        - username
        - password
        - role
        - nama
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          example: "guru_baru_edit"
        password:
          type: string
          minLength: 6
          example: "newpassword"
        role:
          type: string
          enum: [admin, guru, siswa]
          example: "guru"
        nama:
          type: string
          example: "Nama Guru Edit"

    # ── Kelas ──
    Kelas:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nama:
          type: string
          example: "X-A"
      required:
        - id
        - nama
    KelasCreate:
      type: object
      properties:
        nama:
          type: string
          example: "X-A"
      required:
        - nama
    KelasUpdate:
      type: object
      properties:
        nama:
          type: string
          example: "X-B"

    # ── Mata Pelajaran ──
...
    Guru:
      type: object
      properties:
        id:
          type: string
          description: ID Guru (unique)
          example: "G001"
        nama:
          type: string
          example: "Budi Santoso, S.Pd."
        user_id:
          type: integer
          nullable: true
          example: 2
        mengajar_list:
          type: array
          items:
            $ref: '#/components/schemas/GuruMengajar'
      required:
        - id
        - nama
    GuruCreate:
      type: object
      properties:
        id:
          type: string
          description: "ID Guru (auto-generated jika tidak diisi, format: G001)"
          example: "G001"
        nama:
          type: string
          example: "Budi Santoso, S.Pd."
        username:
          type: string
          example: "guru_math"
        password:
          type: string
          example: "password123"
      required:
        - nama
        - username
        - password
    GuruUpdate:
      type: object
      properties:
        nama:
          type: string
          example: "Budi Santoso, M.Pd."

    # ── Guru Mengajar (Junction) ──
    GuruMengajar:
      type: object
      properties:
        id:
          type: integer
          example: 1
        guru_id:
          type: string
          example: "G001"
        guru:
          type: object
          properties:
            id:
              type: string
              example: "G001"
            nama:
              type: string
              example: "Budi Santoso, S.Pd."
        kelas_id:
          type: integer
          example: 1
        kelas:
          type: object
          properties:
            id:
              type: integer
              example: 1
            nama:
              type: string
              example: "X-A"
        mata_pelajaran_id:
          type: integer
          example: 1
        mata_pelajaran:
          type: object
          properties:
            id:
              type: integer
              example: 1
            nama:
              type: string
              example: "Matematika"
      required:
        - id
        - guru_id
        - kelas_id
        - mata_pelajaran_id
        - guru
        - kelas
        - mata_pelajaran
    GuruMengajarCreate:
      type: object
      properties:
        guru_id:
          type: string
          example: "G001"
        kelas_id:
          type: integer
          example: 1
        mata_pelajaran_id:
          type: integer
          example: 1
      required:
        - guru_id
        - kelas_id
        - mata_pelajaran_id

    # ── Nilai ──
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      description: Authenticate user dan return JWT token.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Username atau password salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/change-password:
    put:
      tags:
        - Auth
      summary: Ganti password sendiri
      description: >
        User mengganti password akunnya sendiri.
        Harus menyertakan password lama yang valid.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password berhasil diubah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Password lama salah atau validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Tidak terautentikasi
  /auth/reset-password/{user_id}:
    put:
      tags:
        - Auth
      summary: Admin reset password user
      description: >
        Hanya role Admin.
        Mereset password user lain tanpa perlu password lama.
      operationId: resetPassword
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password berhasil direset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (bukan Admin)
        '404':
          description: User tidak ditemukan

  # ═══════════════════════════════════════════
  # USERS
  # ═══════════════════════════════════════════
  /users:
    get:
      tags:
        - Users
      summary: List semua user
      description: Semua role terautentikasi.
      operationId: listUsers
      responses:
        '200':
          description: Daftar user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          description: Forbidden

    post:
      tags:
        - Users
      summary: Tambah user baru
      description: Hanya role Admin.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User berhasil ditambah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validasi gagal
        '403':
          description: Forbidden
  /users/{user_id}:
    get:
      tags:
        - Users
      summary: Detail user by ID
      description: Semua role terautentikasi.
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Detail user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          description: Forbidden
        '404':
          description: User tidak ditemukan

    put:
      tags:
        - Users
      summary: Update user
      description: Hanya role Admin.
      operationId: updateUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validasi gagal
        '403':
          description: Forbidden
        '404':
          description: User tidak ditemukan

    delete:
      tags:
        - Users
      summary: Hapus user
      description: Hanya role Admin.
      operationId: deleteUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
        '404':
          description: User tidak ditemukan

  # ═══════════════════════════════════════════
  # KELAS
  # ═══════════════════════════════════════════
  /kelas:
    get:
      tags:
        - Kelas
      summary: List semua kelas
      operationId: listKelas
      responses:
        '200':
          description: Daftar kelas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Kelas'

    post:
      tags:
        - Kelas
      summary: Tambah kelas baru
      description: Hanya role Admin.
      operationId: createKelas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KelasCreate'
      responses:
        '201':
          description: Kelas berhasil ditambah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Kelas'
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (bukan Admin)
  /kelas/{id}:
    get:
      tags:
        - Kelas
      summary: Detail kelas by ID
      operationId: getKelasById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Detail kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Kelas'
        '404':
          description: Kelas tidak ditemukan

    put:
      tags:
        - Kelas
      summary: Update data kelas
      description: Hanya role Admin.
      operationId: updateKelas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KelasUpdate'
      responses:
        '200':
          description: Kelas berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Kelas'
        '403':
          description: Forbidden
        '404':
          description: Kelas tidak ditemukan

    delete:
      tags:
        - Kelas
      summary: Hapus kelas
      description: Hanya role Admin.
      operationId: deleteKelas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Kelas berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
        '404':
          description: Kelas tidak ditemukan

  # ═══════════════════════════════════════════
  # MATA PELAJARAN
  # ═══════════════════════════════════════════
...
  /guru:
    get:
      tags:
        - Guru
      summary: List semua guru
      operationId: listGuru
      responses:
        '200':
          description: Daftar guru (termasuk mengajar_list)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Guru'

    post:
      tags:
        - Guru
      summary: Tambah guru baru
      description: >
        Hanya role Admin.
        Otomatis membuat user login untuk guru (role=guru).
        Mata pelajaran diassign via endpoint GuruMengajar.
      operationId: createGuru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuruCreate'
      responses:
        '201':
          description: Guru berhasil ditambah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guru'
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
  /guru/{id}:
    get:
      tags:
        - Guru
      summary: Detail guru by ID
      operationId: getGuruById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "G001"
      responses:
        '200':
          description: Detail guru (termasuk mengajar_list)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guru'
        '404':
          description: Guru tidak ditemukan

    put:
      tags:
        - Guru
      summary: Update data guru
      description: Hanya role Admin.
      operationId: updateGuru
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuruUpdate'
      responses:
        '200':
          description: Guru berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guru'
        '403':
          description: Forbidden
        '404':
          description: Guru tidak ditemukan

    delete:
      tags:
        - Guru
      summary: Hapus guru
      description: >
        Hanya role Admin.
        Juga menghapus user login terkait dan semua assignment mengajar.
      operationId: deleteGuru
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Guru berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
        '404':
          description: Guru tidak ditemukan

  # ═══════════════════════════════════════════
  # GURU MENGAJAR
  # ═══════════════════════════════════════════
  /guru-mengajar:
    get:
      tags:
        - Guru Mengajar
      summary: List semua assignment mengajar
      description: Filterable by guru_id, kelas_id, mata_pelajaran_id.
      operationId: listGuruMengajar
      parameters:
        - name: guru_id
          in: query
          schema:
            type: string
          description: Filter by guru (opsional)
          example: "G001"
        - name: kelas_id
          in: query
          schema:
            type: integer
          description: Filter by kelas (opsional)
          example: 1
        - name: mata_pelajaran_id
          in: query
          schema:
            type: integer
          description: Filter by mata pelajaran (opsional)
          example: 1
      responses:
        '200':
          description: Daftar assignment mengajar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GuruMengajar'

    post:
      tags:
        - Guru Mengajar
      summary: Assign guru ke kelas + mata pelajaran
      description: >
        Hanya role Admin.
        Satu guru bisa mengajar banyak mata pelajaran di banyak kelas.
        Kombinasi (guru_id, kelas_id, mata_pelajaran_id) harus unik.
      operationId: createGuruMengajar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuruMengajarCreate'
      responses:
        '201':
          description: Assignment berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuruMengajar'
        '400':
          description: Validasi gagal (duplikat assignment)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
  /guru-mengajar/{id}:
    get:
      tags:
        - Guru Mengajar
      summary: Detail assignment by ID
      operationId: getGuruMengajarById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Detail assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuruMengajar'
        '404':
          description: Assignment tidak ditemukan

    delete:
      tags:
        - Guru Mengajar
      summary: Hapus assignment mengajar
      description: Hanya role Admin.
      operationId: deleteGuruMengajar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Assignment berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
        '404':
          description: Assignment tidak ditemukan

  # ═══════════════════════════════════════════
  # NILAI
  # ═══════════════════════════════════════════
...